Wrap entry iterator using efficient order.
Iterating over entries using their index may be inefficient as it
generate a lot of cache miss in the internal cluster cache.

Efficient order is entries sorted by they cluster, and so, provides to
most efficient way to use the cluster cache.

`zim-bind` introduce new classes `EntryRangeEfficient` and `IterEfficient`
which wrap zim's `EntryRange` and `iterator` but specialized with
`EntryOrder::efficientOrder`.